great transfer from views to models and utils
