+++ django/test_project2/test_polar/utils.py
@@ -1,26 +1,35 @@
 import tomli
-
+from django.core.cache import cache
+from django.conf import settings
 from django.http import HttpRequest
 
+from test_polar.models import PolarModel
 
+from nosql_adapter import MongoPolar
+
+
+# class Config()
 def _return_configttype() -> list[str]:
     # config = tomli.load(open("../../config.toml", "rb"))
     config = tomli.load(open("config.toml", "rb"))
     return config["running"]["trainingtypes"]
 
 
-def _create_ttype_dict(request: HttpRequest) -> dict:
-    # terribly hacky, but for now it's alright
-    ttype_db = {}
-    ttypes = {"interval", "sprint", "roadrace", "easyrun"}
-    for tt in ttypes:
-        new_val = request.POST["trainingtype-" + tt]
-        if tt != "interval":
-            if new_val == "unknown" or "":
-                new_val = None
-            elif new_val == "false":
-                new_val = False
-            else:
-                new_val = True
-        ttype_db.update({tt: new_val})
-    return ttype_db
+def _set_cache_trainingdata(trainingen: list, cachetime: float) -> None:
+    cache_key = "training_data"
+    cache.clear()
+    cache.set(cache_key, trainingen, cachetime)
+
+
+def get_collections_name() -> list[str]:
+    # config = tomli.load(open("config.toml", "rb"))
+    # database = config["mongodb"]["database"]
+    database = settings.DATABASES["default"]["NAME"]
+    db_table = PolarModel._meta.db_table
+
+    mongpol = MongoPolar(database, db_table)
+    collections = mongpol.getAvailableCollections()
+    collections.sort()
+    collections.remove("__schema__")
+    collections.remove("django_migrations")
+    return collections
