+++ django/test_project2/test_polar/views.py
@@ -13,7 +13,11 @@ from test_polar.models import PolarModel  # , PolarModel_test  # , Testpage
 
 from nosql_adapter import MongoPolar
 
-from .utils import _return_configttype, _create_ttype_dict
+from .utils import (
+    _return_configttype,
+    _set_cache_trainingdata,
+    get_collections_name,
+)  # , _create_ttype_dict,
 
 
 # def _return_configttype() -> list[str]:
@@ -22,38 +26,38 @@ from .utils import _return_configttype, _create_ttype_dict
 #     return config["running"]["trainingtypes"]
 
 
-def _return_trainttype(connection: QuerySet, ttype: str) -> list[Optional[dict]]:
-    if ttype == "sprint":
-        comp = True
-    elif ttype == "easy":
-        ttype = "easyrun"
-        comp = True
-    elif ttype == "road":
-        ttype = "roadrace"
-        comp = True
-    elif ttype == "interval":
-        comp = "interval"
+# def _return_trainttype(connection: QuerySet, ttype: str) -> list[Optional[dict]]:
+#     if ttype == "sprint":
+#         comp = True
+#     elif ttype == "easy":
+#         ttype = "easyrun"
+#         comp = True
+#     elif ttype == "road":
+#         ttype = "roadrace"
+#         comp = True
+#     elif ttype == "interval":
+#         comp = "interval"
 
-    else:
-        print(ttype)
-        return []
+#     else:
+#         print(ttype)
+#         return []
 
-    training = connection.filter(trainingtype={ttype: comp})
-    trainingen = [t for t in training.values()]
-    return trainingen
-
-
-def _return_trainrunning(connection: QuerySet) -> list[Optional[dict]]:
-    training = connection.filter(sport="RUNNING")
-    if len(training) > 0:
-        return [t for t in training.values()]
-    else:
-        return []
+#     training = connection.filter(trainingtype={ttype: comp})
+#     trainingen = [t for t in training.values()]
+#     return trainingen
 
 
-def _return_trainingdata(connection: QuerySet, fname: str) -> dict:
-    trainingen = connection.filter(fname=fname)
-    return trainingen.values()[0]
+# def _return_trainrunning(connection: QuerySet) -> list[Optional[dict]]:
+#     training = connection.filter(sport="RUNNING")
+#     if len(training) > 0:
+#         return [t for t in training.values()]
+#     else:
+#         return []
+
+
+# def _return_trainingdata(connection: QuerySet, fname: str) -> dict:
+#     trainingen = connection.filter(fname=fname)
+#     return trainingen.values()[0]
 
 
 # def _return_lapdata(connection: QuerySet, fname: str) -> list[Optional[dict]]:
@@ -66,16 +70,16 @@ def _return_trainingdata(connection: QuerySet, fname: str) -> dict:
 #         return lapdata
 
 
-def _return_trainingdate(connection: QuerySet, fname: str):
-    training = _return_trainingdata(connection, fname)
-    trainingdate = training["startTime"]
-    return trainingdate
+# def _return_trainingdate(connection: QuerySet, fname: str):
+#     training = _return_trainingdata(connection, fname)
+#     trainingdate = training["startTime"]
+#     return trainingdate
 
 
-def _set_cache_trainingdata(trainingen: list, cachetime: float):
-    cache_key = "training_data"
-    cache.clear()
-    cache.set(cache_key, trainingen, cachetime)
+# def _set_cache_trainingdata(trainingen: list, cachetime: float):
+#     cache_key = "training_data"
+#     cache.clear()
+#     cache.set(cache_key, trainingen, cachetime)
 
 
 # def _create_ttype_dict(request: HttpRequest) -> dict:
@@ -95,44 +99,45 @@ def _set_cache_trainingdata(trainingen: list, cachetime: float):
 #     return ttype_db
 
 
-def _set_database(request: HttpRequest, connection):
-    print("")
-    # print(PolarModel.trainingtype.field.value_from_object("interval"))
-    new_trainingtype = _create_ttype_dict(request)
-    new_description = request.POST["trainingdescription-description"]
-    print(new_description)
-    new_location = request.POST["location"]
+# def _set_database(request: HttpRequest, connection):
+#     print("")
+#     # print(PolarModel.trainingtype.field.value_from_object("interval"))
+#     new_trainingtype = _create_ttype_dict(request)
+#     new_description = request.POST["trainingdescription-description"]
+#     print(new_description)
+#     new_location = request.POST["location"]
 
-    fname = request.POST["fname"]
-    print(fname)
-    training = _return_trainingdata(connection, fname)
-    obj_id = training["_id"]
+#     fname = request.POST["fname"]
+#     print(fname)
+#     training = _return_trainingdata(connection, fname)
+#     obj_id = training["_id"]
 
-    db_table = PolarModel._meta.db_table
-    config = tomli.load(open("config.toml", "rb"))
-    database = config["mongodb"]["database"]
+#     db_table = PolarModel._meta.db_table
+#     config = tomli.load(open("config.toml", "rb"))
+#     database = config["mongodb"]["database"]
 
-    mongpol = MongoPolar(database, db_table)
-    mongpol.updateOne(
-        obj_id,
-        {
-            "location": new_location,
-            "trainingdescription": {"description": new_description},
-            "trainingtype": new_trainingtype,
-        },
-    )
+#     mongpol = MongoPolar(database, db_table)
+#     mongpol.updateOne(
+#         obj_id,
+#         {
+#             "location": new_location,
+#             "trainingdescription": {"description": new_description},
+#             "trainingtype": new_trainingtype,
+#         },
+#     )
 
 
 def select_collections(request: HttpRequest) -> HttpResponse:
     if request.method == "GET":
-        config = tomli.load(open("config.toml", "rb"))
-        database = config["mongodb"]["database"]
-        db_table = PolarModel._meta.db_table
-        mongpol = MongoPolar(database, db_table)
-        collections = mongpol.getAvailableCollections()
-        collections.sort()
-        collections.remove("__schema__")
-        collections.remove("django_migrations")
+        collections = get_collections_name()
+        # config = tomli.load(open("config.toml", "rb"))
+        # database = config["mongodb"]["database"]
+        # db_table = PolarModel._meta.db_table
+        # mongpol = MongoPolar(database, db_table)
+        # collections = mongpol.getAvailableCollections()
+        # collections.sort()
+        # collections.remove("__schema__")
+        # collections.remove("django_migrations")
         return render(request, "home.html", context={"dbtables": collections})
     else:
         return request
@@ -150,22 +155,22 @@ def show_polar(request: HttpRequest) -> HttpResponse:
             print(collection)
             PolarModel.set_dtable(collection)
             print(PolarModel.check())
-        connection = PolarModel.objects.using("default")
+        # connection = PolarModel.objects.using("default")
         # for instance in PolarModel.objects.all():
         #     instance.refresh_from_db()
-        print(connection.all())
+        # print(connection.all())
 
-        trainingen = _return_trainrunning(connection)
+        # trainingen = PolarModel._return_trainrunning()
 
         if "ttypes" not in request.GET:
-            trainingen = _return_trainrunning(connection)
+            trainingen = PolarModel._return_trainrunning()
             # training = connection.filter(sport="RUNNING")
         else:
             ttype = request.GET["ttypes"]
-            print(len(ttype))
-            print(ttype)
+            # print(len(ttype))
+            # print(ttype)
 
-            trainingen = _return_trainttype(connection, ttype)
+            trainingen = PolarModel._return_trainttype(ttype)
             print(len(trainingen))
 
         _set_cache_trainingdata(trainingen, 360)
@@ -185,8 +190,8 @@ def show_polar(request: HttpRequest) -> HttpResponse:
 
 
 def start_adapt(request: HttpRequest) -> HttpResponse:
-    connection = PolarModel.objects.using("default")
-    trainingen = _return_trainrunning(connection)
+    # connection = PolarModel.objects.using("default")
+    trainingen = PolarModel._return_trainrunning()
 
     if request.method == "GET":
         # TODO: add logging
@@ -194,8 +199,9 @@ def start_adapt(request: HttpRequest) -> HttpResponse:
     elif request.method == "POST":
         print(request.POST)
         # print(xx)
-        _set_database(request, connection)
-        trainingen = _return_trainrunning(connection)
+        PolarModel._set_database_adapt(request)
+        # _set_database(request, connection)
+        trainingen = PolarModel._return_trainrunning()
     _set_cache_trainingdata(trainingen, 360)
 
     return render(
@@ -232,11 +238,13 @@ def show_adapt(request: HttpRequest, fname: str):
     if request.method == "GET":
         return request
     elif request.method == "DELETE":
-        data = json.loads(request.body.decode("utf-8"))
-        fname = data.pop("fname", None)
-        connection.filter(fname=fname).delete()
-        trainingen = _return_trainrunning(connection)
+        PolarModel.delete_training(request)
+        # data = json.loads(request.body.decode("utf-8"))
+        # fname = data.pop("fname", None)
+        # connection.filter(fname=fname).delete()
+        trainingen = PolarModel._return_trainrunning()
         _set_cache_trainingdata(trainingen, 360)
+
         return render(
             request,
             "adapt.html",
@@ -252,9 +260,9 @@ def show_adapt(request: HttpRequest, fname: str):
         fname = data.pop("fname", None)
         lapdata = PolarModel.return_lapdata(fname)
         # lapdata = _return_lapdata(connection, fname)
-        ldate = _return_trainingdate(connection, fname)
+        ldate = PolarModel._return_trainingdate(fname)
 
-        training = _return_trainingdata(connection, fname)
+        training = PolarModel._return_trainingdata(fname)
         location = training["location"]
         try:
             description = training["trainingdescription"]["description"]
@@ -310,16 +318,16 @@ def show_adapt(request: HttpRequest, fname: str):
 
 def show_lapdata(request: HttpRequest, fname: str) -> Union[HttpResponse, JsonResponse]:
     if request.method == "GET":
-        print(fname)
+        # print(fname)
 
-        connection = PolarModel.objects.using("default")
+        # connection = PolarModel.objects.using("default")
         print(dir(PolarModel))
         lapdata = PolarModel.return_lapdata(fname)
         print(lapdata)
         # lapdata = _return_lapdata(connection, fname)
         ttypes = _return_configttype()
         # xx
-        ldate = _return_trainingdate(connection, fname)
+        ldate = PolarModel._return_trainingdate(fname)
 
         return render(
             request,
